Implement closing at and next open subtitle in app bar

+++ app/src/main/java/org/cornelldti/flux/data/Facility.kt
@@ -11,8 +11,8 @@ data class Facility(
     // facilityInfo
     var dailyHours: List<Pair<Int, Int>> = listOf(),
     var isOpen: Boolean = false,
-    var nextOpen: Int = -1,
-    var closingAt: Int = -1,
+    var nextOpen: Long = -1L,
+    var closingAt: Long = -1L,
     var campusLocation: CampusLocation = CampusLocation.UNKNOWN,
     // howDense
     var density: Int = -1,
